Move all stories and update import paths
