+++ stories/src/ActionList.stories.tsx
@@ -14,11 +14,11 @@ import {
 import {Meta} from '@storybook/react'
 import React, {forwardRef} from 'react'
 import styled from 'styled-components'
-import {Label, ThemeProvider} from '..'
-import {ActionList as _ActionList} from '../ActionList'
-import {Header} from '../ActionList/Header'
-import BaseStyles from '../BaseStyles'
-import sx from '../sx'
+import {Label, ThemeProvider} from '../../src'
+import {ActionList as _ActionList} from '../../src/ActionList'
+import {Header} from '../../src/ActionList/Header'
+import BaseStyles from '../../src/BaseStyles'
+import sx from '../../src/sx'
 
 const ActionList = Object.assign(_ActionList, {
   Header
