+++ stories/src/ActionList2/fixtures.stories.tsx
@@ -27,14 +27,14 @@ import React, {forwardRef} from 'react'
 import styled from 'styled-components'
 import {DndProvider, useDrag, useDrop} from 'react-dnd'
 import {HTML5Backend} from 'react-dnd-html5-backend'
-import {Label, ThemeProvider} from '../..'
-import {ActionList as _ActionList, ItemProps} from '../../ActionList2'
-import {Header} from '../../ActionList/Header'
-import BaseStyles from '../../BaseStyles'
-import Avatar from '../../Avatar'
-import {ButtonInvisible} from '../../Button'
-import Box from '../../Box'
-import {AnchoredOverlay} from '../../AnchoredOverlay'
+import {Label, ThemeProvider} from '../../../src'
+import {ActionList as _ActionList, ItemProps} from '../../../src/ActionList2'
+import {Header} from '../../../src/ActionList/Header'
+import BaseStyles from '../../../src/BaseStyles'
+import Avatar from '../../../src/Avatar'
+import {ButtonInvisible} from '../../../src/Button'
+import Box from '../../../src/Box'
+import {AnchoredOverlay} from '../../../src/AnchoredOverlay'
 
 const ActionList = Object.assign(_ActionList, {
   Header
