+++ stories/src/ActionMenu2/fixtures.stories.tsx
@@ -1,16 +1,16 @@
 import React from 'react'
 import {Meta} from '@storybook/react'
-import {ThemeProvider} from '../..'
-import BaseStyles from '../../BaseStyles'
-import {ActionMenu} from '../../ActionMenu2'
-import {ActionList} from '../../ActionList2'
-import {Button} from '../../Button2'
-import {IconButton} from '../../Button2/IconButton'
-import Box from '../../Box'
-import Text from '../../Text'
-import TextInput from '../../TextInput'
-import StyledOcticon from '../../StyledOcticon'
-import FormGroup from '../../FormGroup'
+import {ThemeProvider} from '../../../src'
+import BaseStyles from '../../../src/BaseStyles'
+import {ActionMenu} from '../../../src/ActionMenu2'
+import {ActionList} from '../../../src/ActionList2'
+import {Button} from '../../../src/Button2'
+import {IconButton} from '../../../src/Button2/IconButton'
+import Box from '../../../src/Box'
+import Text from '../../../src/Text'
+import TextInput from '../../../src/TextInput'
+import StyledOcticon from '../../../src/StyledOcticon'
+import FormGroup from '../../../src/FormGroup'
 import {
   ServerIcon,
   PlusCircleIcon,
