+++ stories/src/Radio.stories.tsx
@@ -2,8 +2,8 @@ import React, {ChangeEvent, useState} from 'react'
 import {Meta} from '@storybook/react'
 import styled from 'styled-components'
 
-import {BaseStyles, Box, Radio, RadioProps, Text, ThemeProvider} from '..'
-import {get} from '../constants'
+import {BaseStyles, Box, Radio, RadioProps, Text, ThemeProvider} from '../../src'
+import {get} from '../../src/constants'
 
 export default {
   title: 'Forms/Radio',
