+++ stories/src/TextInput.stories.tsx
@@ -1,8 +1,8 @@
 import React, {useState, ReactNode} from 'react'
 import {Meta} from '@storybook/react'
 
-import {BaseStyles, Box, ThemeProvider, Text} from '..'
-import TextInput, {TextInputProps} from '../TextInput'
+import {BaseStyles, Box, ThemeProvider, Text} from '../../src'
+import TextInput, {TextInputProps} from '../../src/TextInput'
 import {CheckIcon} from '@primer/octicons-react'
 
 export default {
